[MLGO][LoopUnroll] Add logging mechanism for loop partial unroll
This is the first part of adding MLGO into LoopUnrollPass. We start
by observing the decision to loop partial unroll.

Loop properties are extracted as input features and the partial unroll
count is the decision. An object MLGOAnalysis will be created in the
first time the LoopPass LoopUnrollPass is triggered, logs will
accumulate and all would be flushed at the end of compilation, in the
dtor of LoopUnrollPass.

Depends on D138368, which introduces LoopPropertiesAnalysis.

Differential Revision: https://reviews.llvm.org/D138369